pyproject.toml
[tool.poetry]
name = "rag-agro"  # substitua pelo nome do seu projeto
version = "0.1.0"
description = ""
authors = ["Lucas Zaia"]


[tool.poetry.dependencies]
python = ">=3.11,<3.13"
beautifulsoup4 = "*"
black = "*"
tiktoken = "*"
openai = "*"
unstructured = "*"
nltk = "*"
fastapi = "*"
jinja2 = "*"
uvicorn = "*"
streamlit = "*"
streamlit-chat = "*"
tqdm = "*"
isort = "*"
langchainhub = "*"
firecrawl-py = "*"
langchain-community = "*"
langsmith = "*"
pinecone = "*"
pillow = "*"
langchain-pinecone = "^0.2.3"

[tool.poetry.dev-dependencies]
